refactored code and test ray docker image
