+++ transforms/language/gneissweb_classification/dpk_gneissweb_classification/__init__.py
@@ -1,3 +1,3 @@
 # SPDX-License-Identifier: Apache-2.0
-from .transform_python import *
+from .runtime import *
 from .transform import *
