+++ transforms/language/gneissweb_classification/dpk_gneissweb_classification/ray/s3.py
@@ -16,7 +16,7 @@ import sys
 
 from data_processing.utils import ParamsUtils
 from data_processing_ray.runtime.ray import RayTransformLauncher
-from dpk_gneissweb_classification.ray.transform import ClassificationRayTransformConfiguration
+from dpk_gneissweb_classification.ray.runtime import ClassificationRayTransformConfiguration
 from dpk_gneissweb_classification.transform import (
     content_column_name_cli_param,
     model_credential_cli_param,
